% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{panel_diff}
\alias{panel_diff}
\title{Compute first difference within panel}
\usage{
panel_diff(x, id)
}
\arguments{
\item{x}{Numeric vector.}

\item{id}{Panel identifier vector.}
}
\value{
Numeric vector with first differences.
}
\description{
Compute first difference within panel
}
\keyword{internal}
