% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{panel_lag}
\alias{panel_lag}
\title{Compute lagged variable within panel}
\usage{
panel_lag(x, id, lag = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{id}{Panel identifier vector.}

\item{lag}{Number of lags.}
}
\value{
Numeric vector with lagged values.
}
\description{
Compute lagged variable within panel
}
\keyword{internal}
