% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtcspqardl.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula for xtcspqardl}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{Formula object}
}
\value{
List with depvar, sr_vars, lr_vars
}
\description{
Parse formula for xtcspqardl
}
\keyword{internal}
