% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_coef_table}
\alias{print_coef_table}
\title{Print coefficient table with significance stars}
\usage{
print_coef_table(tbl, digits = 4, signif.stars = TRUE)
}
\arguments{
\item{tbl}{Data frame with coefficient table.}

\item{digits}{Number of digits.}

\item{signif.stars}{Show significance stars.}
}
\description{
Print coefficient table with significance stars
}
\keyword{internal}
