% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.xtcspqardl}
\alias{residuals.xtcspqardl}
\title{Residuals from xtcspqardl model}
\usage{
\method{residuals}{xtcspqardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtcspqardl"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
This function is not yet implemented for xtcspqardl objects.
}
\description{
Residuals from xtcspqardl model
}
